Refactor and move get_llm, get_env and get_default_thread out
